frontend_eng/requirements.txt
streamlit
pandas
requests
python-dotenv
pypdfium2
reportlab
Pillow
PyPDF2
Authlib